split FileCommunication
